#19 Moved the logic of the context menu to a separate class - ContextMenuLogic
